Antipope
Adds the dreaded antipope, in the form of a role called the 'Doomsayer'. They get EVIL sermons, which inverse the effects of a normal sermon (good for Inhumen, bad for Tennites), can see any Inhumen follower via the Seer trait, inherit the Heretic's torture as the new sole bearer of it, have inverse priest statline, etc.

Just a lot of silly stuff specifically meant to position them as an odd RP class. This entire thing was done as a joke, but I see genuine potential for RP. Is it a good idea? No. Is it funny? Yeah, it is. I'll die on this hill.